Player Health Manager
Create the player health manager in the script called "PlayerHealhAndDeathManager" - which sets the player's health and deals damage when they encounter an enemy. Created a enemy attack system in the "EnemyAttackSystem" script, which deals damage to player when they trigger the box collision on the enemy. Also added a HealthBarSlider into the HUD, which visually communicates when player's health decreases.