Edits to Magic System
Added the bullet system to the other 2 magic types (right click and the combo spell). So what ever mouse click you choose, and bullet represent a visual indicator of where the raycast hits. Also added a destroy to remove the instanciated bullet after a couple of seconds. Also made a few small edits which depletes the mana no matter where the player shoots. Could be at an enemy or at nothing, and the mana system will still deplete.